Move the creation of device queue to assignment
Parse through the json files proved too difficult.  There's
probably an easier way but I wasn't able to figure it out so
I moved the creation of the device threads/queues to the assignment
function.  Basically if the device doesn't exist, create it, it's queue
and then spin up the thread.

It kind of works, but seems significantly slower.  This would make
sense given each thread may be spun up on the fly.  More work is
needed to better fix this issue.